requirements.txt
scikit-learn>=1.4,<2.0
numpy>=1.26,<3.0
python-dotenv>=1.0,<2.0
groq>=0.9,<1.0
